requirements.txt
albumentations==1.4.0
numpy==1.24.4
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
pillow==9.3.0
timm==0.5.4
torch==2.1.0
torchvision==0.16.0
transformers==4.47.0
accelerate==1.6.0
pymupdf==1.26
fastapi
uvicorn
beautifulsoup4
requests
python-multipart